ResourceOpener init
