+++ yandextank/stepper/missile.py
@@ -3,12 +3,11 @@ Missile object and generators
 
 You should update Stepper.status.ammo_count and Stepper.status.loop_count in your custom generators!
 '''
-from util import get_opener
 from itertools import cycle
 from module_exceptions import AmmoFileError
 import info
 import logging
-
+from yandextank.core.tankcore import ResourceOpener
 
 class HttpAmmo(object):
 
@@ -105,7 +104,7 @@ class AmmoFileReader(object):
                     return line
                 chunk_header = line.strip('\r\n')
             return chunk_header
-        opener = get_opener(self.filename)
+        opener = ResourceOpener().get_opener(self.filename)
         with opener(self.filename, 'rb') as ammo_file:
             info.status.af_size = opener.data_length
             # if we got StopIteration here, the file is empty
@@ -148,7 +147,7 @@ class SlowLogReader(object):
         self.filename = filename
 
     def __iter__(self):
-        opener = get_opener(self.filename)
+        opener = ResourceOpener().get_opener(self.filename)
         with opener(self.filename, 'rb') as ammo_file:
             info.status.af_size = opener.data_length
             request = ""
@@ -174,7 +173,7 @@ class LineReader(object):
         self.filename = filename
 
     def __iter__(self):
-        opener = get_opener(self.filename)
+        opener = ResourceOpener().get_opener(self.filename)
         with opener(self.filename, 'rb') as ammo_file:
             info.status.af_size = opener.data_length
             while True:
@@ -204,7 +203,7 @@ class AccessLogReader(object):
         self.log.debug(message)
 
     def __iter__(self):
-        opener = get_opener(self.filename)
+        opener = ResourceOpener().get_opener(self.filename)
         with opener(self.filename, 'rb') as ammo_file:
             info.status.af_size = opener.data_length
             while True:
@@ -247,7 +246,7 @@ class UriReader(object):
         self.log.info("Loading ammo from '%s' using URI format." % filename)
 
     def __iter__(self):
-        opener = get_opener(self.filename)
+        opener = ResourceOpener().get_opener(self.filename)
         with opener(self.filename, 'rb') as ammo_file:
             info.status.af_size = opener.data_length
             while True:
@@ -305,7 +304,7 @@ class UriPostReader(object):
                 else:
                     chunk_header = line.strip('\r\n')
             return chunk_header
-        opener = get_opener(self.filename)
+        opener = ResourceOpener().get_opener(self.filename)
         with opener(self.filename, 'rb') as ammo_file:
             info.status.af_size = opener.data_length
             # if we got StopIteration here, the file is empty
